LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/error-type-2-expected.txt

CONSOLE MESSAGE: SyntaxError: Importing binding name 'default' cannot be resolved by star export entries.
CONSOLE MESSAGE: SyntaxError: Unexpected token '%'

PASS parse error has higher priority than instantiation error
